refactor: 기간 클래스 추가
