+++ src/main/java/nextstep/courses/domain/InvalidPeriodException.java
@@ -1,0 +1,7 @@
+package nextstep.courses.domain;
+
+public class InvalidPeriodException extends RuntimeException {
+    public InvalidPeriodException(String message) {
+        super(message);
+    }
+}
